+++ commands/printer/resetprinter.py
